Working Dealer
The basicDealer class works now yet is missing a few methods to be fully operational. This version of the game uses the old array way of storing a hand's values, this will be phased out with the newer design in later commits. Not recommended to revert to if using newer version unless new version is completely unstable.